add polymorphic source models
Racks now have a nullable reference to a source, which can be from
email, twitter, or seeclickfix currently. If the source is None, that
means that it was submitted through the web. Also, there's a migration
that needs to be run in addition to syncdb, add_rack_source_id.sql, to
add the source_id column to the track table. Also added some convenience
get_absolute_url methods where it made sense for easy url generation to
the source.