update migration script to work with fresh data
one of the tables is created by the initial sql script. i'm commenting
out the creation in the south migration.
we can alternatively decide to update the south migration to execute the
equivalent of the sql script, in which case we won't need to run the sql
script at all.