Fix #205, don't clobber the address when there's errors in the add form.
Problem was that we were hardcoding the map bounds and then
reverse-geocoding the map center to get the address. (Twice!)

As a nice side effect of fixing this, we initialize the map with a
nice readable center point in longitude-latitude, instead of with
cryptic bounds in the EPSG:900913 projection.

As another nice side effect, the map is now centered precisely on the
default marker, rather than slightly off-center.