no longer use live events for pagination
the live events were causing a strange interaction with the map and the filter
form. all clicks were triggering the live event behavior, and a map appeared to
be injecting itself into the boro select. so we just add the pagination
behavior every time we update because it was easier to not use live events for
this case instead of pinpointing exactly what the conflict was