server no longer returns html after voting
this was causing an issue because the html couldn't always be the same
when returned. it had to be different based on what page the request was
coming from. the js now has the smarts to handle the different cases, so
html is no longer necessary to be returned here.

but this commit is just a formality because the js is ignoring the html
server response anyway